Refactor and code cleanup
