+++ example/src/main/scala/geotrellis/server/example/ExampleConf.scala
@@ -1,8 +1,5 @@
 package geotrellis.server.example
 
-import java.net.URI
-
-
 case class ExampleConf(http: ExampleConf.Http, auth: ExampleConf.Auth)
 
 object ExampleConf {
